in-prep for further refactoring, and a 3.x release I've shuffled some things around and gotten test-coverage to 100%.
