+++ lib/parser.js
@@ -1,3 +1,5 @@
+// fancy-pants parsing of argv, originally forked
+// from minimist: https://www.npmjs.com/package/minimist
 var path = require('path');
 var fs = require('fs');
 
@@ -145,8 +147,7 @@ module.exports = function (args, opts) {
                       }
                   });
               } catch (ex) {
-                  console.error('Invalid JSON config file: ' + configPath);
-                  throw ex;
+                  throw Error('invalid json config file: ' + configPath);
               }
             }
         });
@@ -327,8 +328,4 @@ function isNumber (x) {
     if (typeof x === 'number') return true;
     if (/^0x[0-9a-f]+$/i.test(x)) return true;
     return /^[-+]?(?:\d+(?:\.\d*)?|\.\d+)(e[-+]?\d+)?$/.test(x);
-}
-
-function longest (xs) {
-    return Math.max.apply(null, xs.map(function (x) { return x.length }));
 }
