+++ test/fixtures/bin.js
